fix(google-oauth): resolve code review feedback and extract docx util
